# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "仮 API",
        "title": "仮 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "仮",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9800",
    "basePath": "/api",
    "paths": {
        "/rolls/general": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "一般判定の結果を返す",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目標値",
                        "name": "target_value",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "補正値。足すと難しくなり、引くと簡単になる",
                        "name": "ajust",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.GeneralRollOutputPort"
                        },
                        "headers": {
                            "Token": {
                                "type": "string",
                                "description": "qwerty"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponseObject"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ErrorResponseObject": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "ports.GeneralRollOutputPort": {
            "type": "object",
            "properties": {
                "critical": {
                    "type": "boolean"
                },
                "difference": {
                    "type": "integer"
                },
                "fumble": {
                    "type": "boolean"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}